refactor(weather): Move weather services to feature-specific directory
- Relocated all weather-related service classes from lib/services/ to lib/features/weather/services/
to better align with feature-based architecture.
- Updated all corresponding import paths to maintain consistency across layers.
