+++ lib/features/weather/services/i_weather_api_client.dart
@@ -1,5 +1,5 @@
-import '../core/network/response/result.dart';
-import '../features/weather/models/forecast.dart';
+import '../../../core/network/response/result.dart';
+import '../models/forecast.dart';
 
 /// 天気データを取得するAPIクライアントのインターフェース
 abstract class IWeatherApiClient {
