+++ lib/features/weather/services/weather_service_provider.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../features/weather/repositories/firestore_weather_repository_provider.dart';
-import '../features/weather/repositories/weather_repository_provider.dart';
-import '../services/weather_service.dart';
+import '../repositories/firestore_weather_repository_provider.dart';
+import '../repositories/weather_repository_provider.dart';
+import 'weather_service.dart';
 
 final weatherServiceProvider = Provider<WeatherService>((ref) {
   return WeatherService(
